+++ samples/SourceCode.Clay.AspNetCore.Correlation/MvcCorrelationIdSample/MvcCorrelationCustomSample.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
     <TargetFramework>netcoreapp2.2</TargetFramework>
