+++ samples/SourceCode.Clay.AspNetCore.Correlation/MvcCorrelationIdSample/Services/SingletonService.cs
@@ -1,0 +1,16 @@
+using SourceCode.Clay.AspNetCore.Middleware.Correlation;
+
+namespace MvcCorrelationCustomSample.Services
+{
+    public class SingletonService
+    {
+        private readonly ICorrelationContextAccessor _correlationContextAccessor;
+
+        public SingletonService(ICorrelationContextAccessor correlationContextAccessor)
+        {
+            _correlationContextAccessor = correlationContextAccessor;
+        }
+
+        public string GetCorrelationId => _correlationContextAccessor.CorrelationContext.CorrelationId;
+    }
+}
